Data prep, binning, model runs, and plots for a pseudocoral PRYSM run
I apologize for my limited Python knowledge... in an effort to just get something working I binned and plotted the data in R but ran the models in Py. Eventually we'll want to do everything in Py to be consistent but for now at least there's some working model output.