Observed SST coral PSM and modeled vs. observed SST comparisons
I haven't been able to find observational SSS gridded products that go back to 1850... in fact there might not even be any that go back to 1880. I ran the coral psm using observed SST and modeled SSS just for comparison. The modeled and observed SST are quite different, and therefore the resulting pseudo-proxies are quite different.